+++ src/templates/index.php
@@ -8,10 +8,10 @@
  # FILE INFORMATION
  DEFGROUP: Joomla.Template.Site
  INGROUP: MokoCassiopeia
- REPO: https://github.com/mokoconsulting-tech/moko-cassiopeia
- PATH: ./templates/moko-cassiopeia/index.php
+ REPO: https://github.com/mokoconsulting-tech/MokoCassiopeia
+ PATH: ./templates/mokocassiopeia/index.php
  VERSION: 03.06.02
- BRIEF: Main template index file for Moko-Cassiopeia rendering site layout
+ BRIEF: Main template index file for MokoCassiopeia rendering site layout
  */
 
 
@@ -84,7 +84,7 @@ $final     = $pageTitle !== ''
 $this->setTitle($final);
 
 // Template/Media path
-$templatePath = 'media/templates/site/moko-cassiopeia';
+$templatePath = 'media/templates/site/mokocassiopeia';
 
 // Core template CSS
 $wa->useStyle('template.base');   // css/template.css