+++ CHANGELOG.md
@@ -54,7 +54,7 @@
  
 ### Updated
  - Copyright Headers to MokoCodingDefaults standards
- - Fixed `./templates/moko-cassiopeia/index.php` color style injection
+ - Fixed `./templates/mokocassiopeia/index.php` color style injection
  - Upgraded `FontAwesome 6` to `FontAwesome 7 Free`
  - Added `Font Awesome 7 Free` style fallback